requirements.txt
# Core
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.27.0
auto-gptq>=0.7.0
optimum>=1.17.0

# Data
datasets>=2.18.0
huggingface-hub>=0.21.0

# Config
pyyaml>=6.0
omegaconf>=2.3.0

# Analysis & plotting
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
scipy>=1.11.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
